Add ras_concurrent_connections_max metric
I'm not positive this meets the naming standards discussed in the
prometheus documentation here:

https://prometheus.io/docs/practices/naming/

Specifically I have 3 possible issues that spring to mind:

First, it seems like most of the metrics discussed there are
concerned with a single facet (or a bucket of multiple facets), i.e.
`ras_vpn ras_vpn_connections` or `ras_vpn_connections_total`, instead
of `ras_concurrent_connections_max`.

Second, I'm unsure if "concurrent" or "max" is supposed to be a
suffix (or both).

Third, I'm (quite) unsure if qualifiers are supposed to be in the
metric name at all, i.e. should the metric instead be called
"ras_concurrent_connections" with a label for max:

`ras_concurrent_connections{max="123456"}`

With the above in mind, there may be a restructuring of the metrics
format for this project, but my goal is to keep it the same thereafter.